+++ server/src/diagnostics/handlers/coreHandlers.ts
@@ -9,11 +9,9 @@ import RelateTagHandler from './relateHandler';
 import ShorthandModifierHandler from './shorthandModifierHandler';
 import StatamicVersionHandler from './statamicVersionHandler';
 import TagsThatErrorHandler from './tagsThatErrorHandler';
-import UrlInputHandler from './urlInputHandler';
 
 const CoreHandlers: IDiagnosticsHandler[] = [
     DataDumpHandler,
-    UrlInputHandler,
     MixedModifierHandler,
     ModifierRuntimeTypeHandler,
     DoubleColonHandler,