+++ server/src/test/formatter_tags.test.ts
@@ -90,6 +90,7 @@ After Partial
             {{ slot:test }}
                 NameStart{{ value }}NameEnd
             {{ /slot:test }}
+
             Normal Slot Content ({{ value }})
         {{ /partial:nested }}
         After Partial